feat(db)!: migrate authentication schema
Flatten user roles, replace refresh token storage, align movement timestamps and recreate dependent report views.

BREAKING CHANGE: invalidate existing refresh sessions and replace legacy user role tables.

+++ apps/backend/src/main/java/io/github/artsobol/kurkod/feature/report/dto/response/ChickensByWorkshopAndBreedResponse.java
@@ -1,4 +1,4 @@
 package io.github.artsobol.kurkod.feature.report.dto.response;
 
 public record ChickensByWorkshopAndBreedResponse(
-    Long workshopId, Long workshopNumber, Long breedId, String breedName, Long chickensCount) {}
+    Long workshopId, Integer workshopNumber, Long breedId, String breedName, Long chickensCount) {}
